% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Checks for whole numbers}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a logical vector indicating whether the element is a whole number
}
\description{
Checks for whole numbers
}
\seealso{
\link{\code{permutations}}, examples in \link{\code{is.integer}}
}
