% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{missing}
\alias{missing}
\title{Determines missing tiles in a challenge}
\usage{
missing(n)
}
\arguments{
\item{n, }{integer (1-64) indicating the challenge number}
}
\value{
a table of counts for missing tiles in a consistent format
}
\description{
Calculates which tiles are missing from a given challenge.
}
\examples{
missing(1)
missing(64)

}
