% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{Number of tile permutations}
\usage{
permutations(n)
}
\arguments{
\item{n}{a count vector indicating the number each unique tiles}
}
\value{
the number of unique permutations
}
\description{
Marble circuit is a puzzle game containing 10 tiles, 4 unique types, that
must be placed in a particular order to obtain a desired number of marbles in
each of five target spaces. There are 4 unique types of tile.
This function will determine the number of unique arrangements
of a subset of these tiles. The user must indicate the number of each type
of tile.
}
\examples{
permutations(1)
permutations(rep(1,4))
permutations(c(3,3,2,2))

}
\author{
Jarad Niemi
}
