% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targets}
\alias{targets}
\title{The count of marbles in each target}
\format{
A data frame with 640 rows and 4 variables:
\describe{
\item{type}{type of card: Challegne or Solution}
\item{number}{challenge number (1-64)}
\item{target}{target number (1-5), numbered from left to right}
\item{count}{the number of marbles in that target}
}
}
\usage{
targets
}
\description{
Marble circuit has 8 marbles that start from 8 different gates and run
through the tiles to 5 targets. This provides the count of marbles for each
target, numbered 1 to 5 from left to right.
}
\keyword{datasets}
